Move argument validation and parsing logic to its own file
